chore: move assert functions to general
