+++ packages/shared/src/index.ts
@@ -1,4 +1,3 @@
 export * from './env'
 export * from './general'
 export * from './constants'
-export * from './assert'
